runtime: use wrappers for op= map operations
Add mapassignop* functions for op= operations so we clear reused
values and don't need to zero during all map deletes.

In this case mapassign* should return boolean flag if value was reused.
It adds overhead to mapassing operations.

benchstat ~/hackathon8/mapassign_before_op.txt ~/hackathon8/mapassign_op_reused.txt
name                             old time/op  new time/op  delta
MapAssign/Int32/256-8            14.2ns ±10%  15.0ns ± 8%   +5.54%  (p=0.000 n=18+20)
MapAssign/Int32/65536-8          29.4ns ± 2%  33.4ns ±17%  +13.59%  (p=0.000 n=18+20)
MapAssign/Int64/256-8            13.2ns ± 6%  13.5ns ± 8%     ~     (p=0.207 n=20+20)
MapAssign/Int64/65536-8          28.1ns ± 2%  29.8ns ± 4%   +6.03%  (p=0.000 n=19+20)
MapAssign/Str/256-8              15.1ns ± 6%  15.4ns ± 7%   +2.36%  (p=0.005 n=20+20)
MapAssign/Str/65536-8            33.3ns ± 3%  36.7ns ± 9%  +10.01%  (p=0.000 n=18+20)
MapOperatorAssign/Int32/256-8    12.8ns ± 4%  16.0ns ± 5%  +24.44%  (p=0.000 n=19+19)
MapOperatorAssign/Int32/65536-8  27.9ns ± 5%  32.8ns ± 3%  +17.28%  (p=0.000 n=20+19)
MapOperatorAssign/Int64/256-8    13.0ns ± 5%  15.1ns ±13%  +16.27%  (p=0.000 n=20+20)
MapOperatorAssign/Int64/65536-8  27.8ns ± 4%  30.5ns ± 7%   +9.81%  (p=0.000 n=19+19)
MapOperatorAssign/Str/256-8      1.65µs ± 3%  1.56µs ± 9%   -5.36%  (p=0.000 n=18+20)
MapOperatorAssign/Str/65536-8     327ns ± 8%   233ns ± 9%  -28.61%  (p=0.000 n=16+16)

Fixes #26059

+++ src/runtime/map_test.go
@@ -1066,7 +1066,7 @@ func TestIncrementAfterDeleteValueInt32(t *testing.T) {
 	const key1 = 12
 	const key2 = 13
 
-	m := make(map[int]int32)
+	m := make(map[int32]int32)
 	m[key1] = 99
 	delete(m, key1)
 	m[key2]++
@@ -1079,7 +1079,7 @@ func TestIncrementAfterDeleteValueInt64(t *testing.T) {
 	const key1 = 12
 	const key2 = 13
 
-	m := make(map[int]int64)
+	m := make(map[int64]int64)
 	m[key1] = 99
 	delete(m, key1)
 	m[key2]++
